Moved XSLog classes to separate module
